requirements.txt
flask==2.3.2
pandas==2.0.3
scikit-learn==1.3.0
pyyaml==6.0.1
joblib==1.3.2